+++ packages/widget/src/types/widget.ts
@@ -18,6 +18,7 @@ import type {
 import type { TypographyOptions } from '@mui/material/styles/createTypography';
 import type { CSSProperties, ReactNode, RefObject } from 'react';
 import type { LanguageKey, LanguageResources } from '../providers';
+import type { SplitSubvariantOptions } from '../stores';
 
 export type WidgetVariant = 'default' | 'expandable' | 'drawer';
 
@@ -111,8 +112,9 @@ export interface WidgetConfig {
   contractCompactComponent?: ReactNode;
   contractTool?: WidgetContractTool;
 
-  fee?: number;
   integrator: string;
+  apiKey?: string;
+  fee?: number;
   referrer?: string;
 
   routePriority?: Order;
@@ -121,6 +123,7 @@ export interface WidgetConfig {
 
   variant?: WidgetVariant;
   subvariant?: WidgetSubvariant;
+  subvariantOptions?: SplitSubvariantOptions;
 
   appearance?: Appearance;
   theme?: ThemeConfig;